write module
